+++ README.md
@@ -2,6 +2,14 @@
 
 1. HOOKS //
 
+현재까지의 방식: 매번 반복하며 보지않고 처음부터 끝까지 코딩이 가능할 정도의 이해도를 요하여 반복하며 React의 전반적인 이해도를 위해 comment를 통해 질의응답하며 진행함.
+study, study1 -> 2 ... 식으로 브랜치를 나누며 진행하고 master 브랜치에 push함.
+모르는것, 이해가 안가는점들을 혼자서 찾아나가는 식으로 학습.
+
+차후 방식의 개선 혹은 변경점이 있으시면 코멘트 부탁드립니다.
+아무래도 미아님으로 바뀌었으니 학습방식의 변화가 필요할 시점일것 같습니다.
+그리고 React HOOK강의를 듣기위해 내일 하루 PR 보류하겠습니다!
+
 <!-- 1. return, render
 1. class, component
 1. state, props
