--- src/routes/About.js
@@ -1,16 +1,0 @@
-import React from "react";
-import "./About.css";
-
-function About() {
-  return (
-    <div className="about__container">
-      <span>
-        "일단 뭉탱이로 있어라. 그렇다면 당신이 유링게슝해도 아이그냥 있을
-        것이다."
-      </span>
-      <span>- 코현실주의 화가 'Mungtaeng-i Van Kane' -</span>
-    </div>
-  );
-}
-
-export default About;
